Refactored classes and packages to match new groupId
