+++ src/backend/oracle-data-interface/src/main/java/ca/bc/gov/open/jag/tco/oracledatainterface/model/Dispute.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.trafficcourtsonline.oracledatainterface.model;
+package ca.bc.gov.open.jag.tco.oracledatainterface.model;
 
 import java.sql.Date;
 import java.util.List;
