+++ src/backend/oracle-data-interface/src/main/java/ca/bc/gov/open/jag/tco/oracledatainterface/repository/DisputeRepository.java
@@ -1,8 +1,8 @@
-package ca.bc.gov.trafficcourtsonline.oracledatainterface.repository;
+package ca.bc.gov.open.jag.tco.oracledatainterface.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import ca.bc.gov.trafficcourtsonline.oracledatainterface.model.Dispute;
+import ca.bc.gov.open.jag.tco.oracledatainterface.model.Dispute;
 
 public interface DisputeRepository extends CrudRepository<Dispute, Integer>{
