+++ src/backend/oracle-data-interface/src/main/java/ca/bc/gov/open/jag/tco/oracledatainterface/service/DisputeService.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.trafficcourtsonline.oracledatainterface.service;
+package ca.bc.gov.open.jag.tco.oracledatainterface.service;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,8 +6,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import ca.bc.gov.trafficcourtsonline.oracledatainterface.model.Dispute;
-import ca.bc.gov.trafficcourtsonline.oracledatainterface.repository.DisputeRepository;
+import ca.bc.gov.open.jag.tco.oracledatainterface.model.Dispute;
+import ca.bc.gov.open.jag.tco.oracledatainterface.repository.DisputeRepository;
 
 @Service
 public class DisputeService {
